refactor: move `to_definitions` dependencies to `definitions.rs`
